modificando filter.c: rb e wb; modificando helpers.c convertendo media para o tipo certo

+++ week-4/filter-less/helpers.c
@@ -8,7 +8,7 @@ void grayscale(int height, int width, RGBTRIPLE image[height][width])
     {
         for (int j = 0; j < width; j++)
         {
-            BYTE media = round((image[i][j].rgbtBlue + image[i][j].rgbtGreen + image[i][j].rgbtRed)/3.0);    
+            BYTE media = (BYTE) round((image[i][j].rgbtBlue + image[i][j].rgbtGreen + image[i][j].rgbtRed)/3.0);    
             image[i][j].rgbtBlue = media;
             image[i][j].rgbtGreen = media;
             image[i][j].rgbtRed = media;
